Fixed Vase, Bowls and Shelves; Changed Mass
Fixed Vase and bowls near door that fell as soon as the game starts. Changed mass of all china so the lighter china falls off sooner than heavier ones. Shelves with Plates on the very top and RoundTable are 'isKinematic'. Plate next to Vase next to door is also 'isKinematic'